Sithafal-Task-2-Chat-with-Website-Using-RAG-Pipeline-main/Sithafal-Task-2-Chat-with-Website-Using-RAG-Pipeline-main/Sithafal_Task_2/requirements.txt
requests
beautifulsoup4
sentence-transformers
faiss-cpu
transformers
numpy
